+++ erc1191/src/main/kotlin/org/kethereum/erc1191/ERC1191.kt
@@ -1,4 +1,4 @@
-package org.kethereum.eip1191
+package org.kethereum.erc1191
 
 import org.kethereum.erc55.isValid
 import org.kethereum.keccakshortcut.keccak
@@ -11,7 +11,7 @@ import java.util.*
 EIP-1191 Checksum as in https://github.com/ethereum/EIPs/blob/master/EIPS/eip-1191.md
  */
 
-fun Address.withEIP1191Checksum(chainId: ChainId) = "${chainId.value}${hex}".toLowerCase(Locale.ROOT).toByteArray().keccak().toNoPrefixHexString().let { hexHash ->
+fun Address.withERC1191Checksum(chainId: ChainId) = "${chainId.value}${hex}".toLowerCase(Locale.ROOT).toByteArray().keccak().toNoPrefixHexString().let { hexHash ->
     Address(cleanHex.mapIndexed { index, hexChar ->
         when {
             hexChar in '0'..'9' -> hexChar
@@ -21,11 +21,11 @@ fun Address.withEIP1191Checksum(chainId: ChainId) = "${chainId.value}${hex}".toL
     }.joinToString(""))
 }
 
-private fun Address.hasValidEIP1191ChecksumAssumingValidAddress(chainId: ChainId) = withEIP1191Checksum(chainId).hex == hex
+private fun Address.hasValidERC1191ChecksumAssumingValidAddress(chainId: ChainId) = withERC1191Checksum(chainId).hex == hex
 
-fun Address.hasValidEIP1191Checksum(chainId: ChainId) = isValid() && hasValidEIP1191ChecksumAssumingValidAddress(chainId)
-fun Address.hasValidEIP1191ChecksumOrNoChecksum(chainId: ChainId) = isValid() &&
-        (hasValidEIP1191ChecksumAssumingValidAddress(chainId) ||
+fun Address.hasValidERC1191Checksum(chainId: ChainId) = isValid() && hasValidERC1191ChecksumAssumingValidAddress(chainId)
+fun Address.hasValidERC1191ChecksumOrNoChecksum(chainId: ChainId) = isValid() &&
+        (hasValidERC1191ChecksumAssumingValidAddress(chainId) ||
                 cleanHex.toLowerCase() == cleanHex ||
                 cleanHex.toUpperCase() == cleanHex)
