+++ settings.gradle.kts
@@ -52,4 +52,4 @@ include(":types")
 include(":uri_common")
 include(":userdoc")
 include(":wallet")
-include(":eip1191")
+include(":erc1191")
